+++ docs/src/components/Layout/Sidebar.tsx
@@ -61,7 +61,8 @@ const NavLink = ({
     query: { platform = 'react' },
     pathname,
   } = useCustomRouter();
-  const isCurrent = pathname === `/[platform]${href}`;
+
+  const isCurrent = pathname.includes(href);
   const classNames = `${
     tertiary ? 'docs-tertiary-nav-link' : 'docs-secondary-nav-link'
   } ${isCurrent ? 'current' : ''}`;
@@ -102,7 +103,7 @@ const NavLinkComponentsSection = ({ heading, components, ...props }) => {
         textTransform="uppercase"
         letterSpacing="0.125em"
         color={tokens.colors.font.tertiary}
-        padding={`${tokens.space.small} ${tokens.space.medium} ${tokens.space.xs} var(--secondary-nav-indent)`}
+        padding={`${tokens.space.large} ${tokens.space.medium} ${tokens.space.xs} var(--secondary-nav-indent)`}
       >
         {heading}
       </Text>