+++ docs/src/pages/[platform]/components/radiogroupfield/examples/RadioGroupFieldThemeExample.tsx
@@ -1,32 +1,22 @@
-import {
-  Radio,
-  RadioGroupField,
-  ThemeProvider,
-  Theme,
-} from '@aws-amplify/ui-react';
+import { Radio, RadioGroupField, ThemeProvider } from '@aws-amplify/ui-react';
 
-const theme: Theme = {
-  name: 'radio-theme',
+const theme = {
+  name: 'radiogroup-theme',
   tokens: {
-    colors: {
-      border: {
-        // Radio focus state inherits this global focus color
-        focus: { value: '{colors.brand.secondary.80}' },
-      },
-      font: {
-        primary: { value: '{colors.brand.secondary.60}' },
-        secondary: { value: '{colors.brand.secondary.40}' },
-      },
-    },
     components: {
-      radio: {
-        button: {
+      radiogroup: {
+        label: {
+          color: { value: '{colors.blue.80}' },
+        },
+        radio: {
           borderWidth: { value: '{borderWidths.small}' },
-          borderColor: { value: '{colors.brand.secondary.60}' },
-          backgroundColor: { value: '{colors.brand.secondary.10}' },
-          color: { value: '{colors.brand.secondary.10}' },
+          borderColor: { value: '{colors.blue.60}' },
+          backgroundColor: { value: '{colors.blue.20}' },
           _checked: {
-            color: { value: '{colors.brand.secondary.80}' },
+            color: { value: '{colors.blue.80}' },
+          },
+          label: {
+            color: { value: '{colors.blue.80}' },
           },
         },
       },
@@ -34,16 +24,12 @@ const theme: Theme = {
   },
 };
 
-const options = ['html', 'css', 'javascript'];
-
 export const RadioGroupFieldThemeExample = () => (
   <ThemeProvider theme={theme} colorMode="light">
-    <RadioGroupField label="Language" name="language6" defaultValue="html">
-      {options.map((option) => (
-        <Radio key={option} value={option}>
-          {option}
-        </Radio>
-      ))}
+    <RadioGroupField label="Themed example" name="themed" defaultValue="blue">
+      <Radio value="blue">Blue</Radio>
+      <Radio value="navy">Navy</Radio>
+      <Radio value="cerulean">Cerulean</Radio>
     </RadioGroupField>
   </ThemeProvider>
 );