+++ packages/react/src/primitives/types/field.ts
@@ -28,7 +28,8 @@ export interface FieldProps {
   /**
    * @description
    * Visually hide label (not recommended in most cases)
-   * @default false
+   * @default
+   * false
    */
   labelHidden?: boolean;
 }